+++ src/core/utils/index-script.ts
@@ -4,7 +4,7 @@ export { debounce } from './debounce.js';
 export { randomId } from './id.js';
 export { escapeHtml, numDelim } from './html.js';
 export type { TypedArray } from './is-type.js';
-export { isTypedArray, isArray, isSet, isRegExp } from './is-type.js';
+export { isTypedArray, isArray, isSet, isRegExp, isError } from './is-type.js';
 export { jsonStringifyAsJavaScript, jsonStringifyInfo } from './json.js';
 export type * from './load-data.js';
 export * from './load-data.js';